Changed format of zoneddate time.
This was a bit weird because the definition in the json file is java.time.zoneddatetime. This is pretty clearly an object:https://docs.oracle.com/javase/8/docs/api/java/time/ZonedDateTime.html
But in all the json we get back the value of these fields are longs.

So I just defined it as a long.