Update generated Go client to accept Context (#50)
* Update Go client to accept Context

Updates generated Go client to allow context.Context to optionally be provided by overloading functions to have WithContext equivalent with context.Context as the first parameter.

* Update to use context.TODO()

Use context.TODO() instead of context.Background() based on Go's recommendation. Code should use context.TODO when it's unclear which Context to use or it is not yet available (because the surrounding function has not yet been extended to accept a Context parameter).